requirements.txt
cffi>=1.14.0
cmake>=3.17.2
ctypes>=1.0.2
cycler>=0.10.0
furture>=0.17.1
joblib>=0.14.1
kiwisolver>=1.2.0
matplotlib>=3.2.1
mkl-include>=2019.0
mkl>=2019.0
ninja>=1.7.2
numpy>=1.15.3
pillow>=5.3.0
pyparsing>=2.4.7
python-dateutil>=2.8.1
pyyaml>=3.13
scikit-learn>=0.22.2.post1
setuptools>=46.1.3
six>=1.14.0
sklearn>=0.0
torch>=1.5.0
torchvision>=0.6.0